requirements.txt.txt
altair==5.0.1
atlastk==0.13.2
attrs==23.1.0
audioread==3.0.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
filelock==3.12.2
fonttools==4.42.0
gitdb==4.0.10
GitPython==3.1.32
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.40.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.5
networkx==3.1
numba==0.57.1
numpy==1.24.4
packaging==23.1
pandas==2.0.3
patsy==0.5.3
Pillow==9.5.0
platformdirs==3.10.0
plotly==5.16.1
plotly-express==0.4.1
pooch==1.7.0
protobuf==4.24.0
pyarrow==12.0.1
PyAudio==0.2.13
pycparser==2.21
pydeck==0.8.0
pydub==0.25.1
Pygments==2.16.1
Pympler==1.0.1
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
referencing==0.30.2
requests==2.31.0
rich==13.5.2
rpds-py==0.9.2
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
smmap==5.0.0
soundfile==0.12.1
soxr==0.3.6
statsmodels==0.14.0
streamlit==1.25.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
tornado==6.3.3
tortoise==0.1.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
validators==0.21.2
watchdog==3.0.0
zipp==3.16.2
